move some functions and data to separated files
